+++ include/GafferScene/ScenePlug.h
@@ -66,13 +66,20 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		/// Only accepts ScenePlug inputs.
 		bool acceptsInput( const Gaffer::Plug *input ) const override;
 
-		/// @name Child plugs
-		/// Different aspects of the scene are passed through different
-		/// child plugs.
-		////////////////////////////////////////////////////////////////////
-		//@{
+		/// Child plugs
+		/// ===========
+		///
+		/// Different properties of the scene are represented by different child
+		/// plugs of the ScenePlug.
+
+		/// Location properties
+		/// -------------------
+		///
+		/// These plugs require the `scenePathContextName` variable to
+		/// be provided by the current context.
+		///
 		/// The plug used to pass the bounding box of the current location in
-		/// the scene graph. The bounding box is supplied /without/ the
+		/// the scene graph. The bounding box is supplied _without_ the
 		/// transform applied.
 		Gaffer::AtomicBox3fPlug *boundPlug();
 		const Gaffer::AtomicBox3fPlug *boundPlug() const;
@@ -89,6 +96,21 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		/// location in the scene graph.
 		Gaffer::InternedStringVectorDataPlug *childNamesPlug();
 		const Gaffer::InternedStringVectorDataPlug *childNamesPlug() const;
+		/// Represents the existence of the current location. Value is
+		/// `true` if the location exists and `false` otherwise. This is computed
+		/// automatically by querying `childNamesPlug()` at all ancestor locations,
+		/// but with significantly better performance than is achievable directly.
+		Gaffer::BoolPlug *existsPlug();
+		const Gaffer::BoolPlug *existsPlug() const;
+		/// Provides the union of the bounding boxes of all the children
+		/// of the current location, transformed using their respective
+		/// transforms.
+		Gaffer::AtomicBox3fPlug *childBoundsPlug();
+		const Gaffer::AtomicBox3fPlug *childBoundsPlug() const;
+
+		/// Global properties
+		/// -----------------
+		///
 		/// The plug used to pass renderer options including output etc,
 		/// represented as a CompoundObject. Note that this is not sensitive
 		/// to the "scene:path" context entry.
@@ -106,16 +128,16 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		/// which set to compute.
 		Gaffer::PathMatcherDataPlug *setPlug();
 		const Gaffer::PathMatcherDataPlug *setPlug() const;
-		//@}
 
-		/// @name Context management
+		/// Context management
+		/// ==================
+		///
 		/// The child Plugs are expected to be evaluated in the context
 		/// of a particular location in the scenegraph, so that the
 		/// scenegraph can be evaluated piecemeal, rather than all needing
 		/// to exist at once. These members provide utilities for
 		/// constructing relevant contexts.
-		////////////////////////////////////////////////////////////////////
-		//@{
+
 		/// The type used to specify the current scene path in
 		/// a Context object.
 		typedef std::vector<IECore::InternedString> ScenePath;
@@ -174,9 +196,10 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 			/// ThreadState documentation for more details.
 			GlobalScope( const Gaffer::ThreadState &threadState );
 		};
-		//@}
 
-		/// @name Convenience accessors
+		/// Convenience accessors
+		/// =====================
+		///
 		/// These functions create temporary Contexts specifying the necessary
 		/// variables and then return the result of calling getValue() or hash()
 		/// on the appropriate child plug. Note that if you wish to evaluate
@@ -185,10 +208,9 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		/// getValue() directly.
 		///
 		/// > Note : It is a programming error to trigger a compute for a
-		/// > location which does not exist. Use the `exists()` method to
-		/// > verify existence where necessary.
-		////////////////////////////////////////////////////////////////////
-		//@{
+		/// > location which does not exist. Use the `existsPlug()` and/or
+		/// > the `exists()` method to verify existence where necessary.
+
 		/// Returns the bound for the specified location.
 		Imath::Box3f bound( const ScenePath &scenePath ) const;
 		/// Returns the local transform at the specified scene path.
@@ -201,6 +223,11 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		IECore::CompoundObjectPtr fullAttributes( const ScenePath &scenePath ) const;
 		IECore::ConstObjectPtr object( const ScenePath &scenePath ) const;
 		IECore::ConstInternedStringVectorDataPtr childNames( const ScenePath &scenePath ) const;
+		/// Returns true if the specified location exists.
+		bool exists( const ScenePath &scenePath ) const;
+		/// Returns the union of the bounding boxes of all the children
+		/// of `scenePath`, transformed using their respective transforms.
+		Imath::Box3f childBounds( const ScenePath &scenePath ) const;
 		/// Prefer this to bare `globalsPlug()->getValue()` calls when
 		/// accessing globals from within a per-location computation. It
 		/// uses GlobalScope to remove unnecessary context variables which
@@ -220,29 +247,15 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		IECore::MurmurHash fullAttributesHash( const ScenePath &scenePath ) const;
 		IECore::MurmurHash objectHash( const ScenePath &scenePath ) const;
 		IECore::MurmurHash childNamesHash( const ScenePath &scenePath ) const;
+		IECore::MurmurHash childBoundsHash( const ScenePath &scenePath ) const;
 		/// See comments for `globals()` method.
 		IECore::MurmurHash globalsHash() const;
 		/// See comments for `setNames()` method.
 		IECore::MurmurHash setNamesHash() const;
 		IECore::MurmurHash setHash( const IECore::InternedString &setName ) const;
-		//@}
 
-		/// Returns true if the specified location exists. This is achieved
-		/// by querying `childNames()` at all ancestor locations, but with
-		/// significantly better performance than is achievable via the public
-		/// API alone.
-		bool exists( const ScenePath &scenePath ) const;
-		/// As above, but for the location specified by the current context.
-		bool exists() const;
-
-		/// Returns the union of the bounding boxes of all the children
-		/// of `scenePath`, transformed using their respective transforms.
-		Imath::Box3f childBounds( const ScenePath &scenePath ) const;
-		/// As above, but for the location specified by the current context.
-		Imath::Box3f childBounds() const;
-		/// Returns a hash representing the results of `childBounds()`.
-		IECore::MurmurHash childBoundsHash( const ScenePath &scenePath ) const;
-		IECore::MurmurHash childBoundsHash() const;
+		/// Utility methods
+		/// ===============
 
 		/// Utility function to convert a string into a path by splitting on '/'.
 		/// \todo Many of the places we use this, it would be preferable if the source data was already
@@ -250,22 +263,18 @@ class GAFFERSCENE_API ScenePlug : public Gaffer::ValuePlug
 		static void stringToPath( const std::string &s, ScenePlug::ScenePath &path );
 		static void pathToString( const ScenePlug::ScenePath &path, std::string &s );
 
+		/// Deprecated methods
+		/// ==================
+
+		/// \deprecated Use `existsPlug()->getValue()` instead.
+		bool exists() const;
+
 	private :
 
-		// Private plugs that are used to implement the `exists()` method.
-		// Values for these are computed automatically by SceneNode, hence
-		// the friendship.
-
-		friend class SceneNode;
-
-		Gaffer::BoolPlug *existsPlug();
-		const Gaffer::BoolPlug *existsPlug() const;
-
+		// Private plug used for the computation of `existsPlug()` by SceneNode.
 		Gaffer::InternedStringVectorDataPlug *sortedChildNamesPlug();
 		const Gaffer::InternedStringVectorDataPlug *sortedChildNamesPlug() const;
-
-		Gaffer::AtomicBox3fPlug *childBoundsPlug();
-		const Gaffer::AtomicBox3fPlug *childBoundsPlug() const;
+		friend class SceneNode;
 
 };