+++ src/GafferSceneTest/ContextSanitiser.cpp
@@ -65,9 +65,7 @@ namespace
 {
 
 const InternedString g_internalOut( "__internalOut" );
-const InternedString g_exists( "__exists" );
 const InternedString g_sortedChildNames( "__sortedChildNames" );
-const InternedString g_childBounds( "__childBounds" );
 
 } // namespace
 
@@ -98,11 +96,11 @@ void ContextSanitiser::processStarted( const Gaffer::Process *process )
 			process->plug() != scene->attributesPlug() &&
 			process->plug() != scene->objectPlug() &&
 			process->plug() != scene->childNamesPlug() &&
-			// Private plugs, so we have no choice but to test
-			// for them by name.
-			process->plug()->getName() != g_exists &&
-			process->plug()->getName() != g_sortedChildNames &&
-			process->plug()->getName() != g_childBounds
+			process->plug() != scene->existsPlug() &&
+			process->plug() != scene->childBoundsPlug() &&
+			// Private plug, so we have no choice but to test
+			// for it by name.
+			process->plug()->getName() != g_sortedChildNames
 		)
 		{
 			if( process->context()->get<IECore::Data>( ScenePlug::scenePathContextName, nullptr ) )